Finished backend api, has full CRUD functionality

+++ backend/controllers/postController.js
@@ -3,12 +3,13 @@
 const asyncHandler = require('express-async-handler')
 
 const Post = require('../models/postModel')
+const User = require('../models/userModel')
 
 // @desc Get Posts
 // @route GET /api/posts
 // @access Private
 const getPosts = asyncHandler(async (req, res) => {
-    const posts = await Post.find()
+    const posts = await Post.find({ user: req.user.id})
 
     res.status(200).json(posts)
 })
@@ -23,7 +24,8 @@ const setPost = asyncHandler(async (req, res) => {
     }
 
     const post = await Post.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json(post)
@@ -41,6 +43,20 @@ const updatePost = asyncHandler(async (req, res) => {
         throw new Error('Post not found')
     }
 
+    const user = await User.findById(req.user.id)
+
+    // check if user exists
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // make sure the logged in user matches the post user
+    if(post.user.toString() != user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     const updatedPost = await Post.findByIdAndUpdate(req.params.id, req.body, {new: true})
 
     res.status(200).json(updatedPost)
@@ -56,6 +72,20 @@ const deletePost = asyncHandler(async (req, res) => {
     if(!post){
         res.status(400)
         throw new Error('Post not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    // check if user exists
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // make sure the logged in user matches the post user
+    if(post.user.toString() != user.id){
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     await post.deleteOne()
